fixed bug where size of large groups wasn't computed correctly

+++ src/main/kotlin/ThreeRubiksCube.kt
@@ -1,6 +1,7 @@
 package org.example
 
 import org.example.Permutation.Companion.E
+import java.math.BigInteger
 
 class ThreeRubiksCube {
 
@@ -37,7 +38,7 @@ class ThreeRubiksCube {
     val generators : List<Permutation>
         get() = dictionary.keys.toList()
 
-    val size : Long
+    val size : BigInteger
         get() = SchreierSims(generators).size
 
     fun stabilizersThatPermute(ps: List<Int>): List<Permutation> {
